requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
soupsieve==2.5

requests~=2.31.0
google~=3.0.0
pandas~=2.2.0
numpy~=1.26.3
gensim~=4.3.2
nltk~=3.8.1
sumy~=0.11.0
openai~=0.28.0
python-dotenv~=1.0.1